Move NodeRecord to primitives
along with NodeKey and Octets
