+++ crates/net/discv4/src/config.rs
@@ -3,9 +3,9 @@
 //! This basis of this file has been taken from the discv5 codebase:
 //! https://github.com/sigp/discv5
 
-use crate::node::NodeRecord;
 use bytes::{Bytes, BytesMut};
 use reth_net_common::ban_list::BanList;
+use reth_primitives::NodeRecord;
 use reth_rlp::Encodable;
 use std::{
     collections::{HashMap, HashSet},
