+++ crates/primitives/Cargo.toml
@@ -14,9 +14,11 @@ reth-rlp = { path = "../common/rlp", features = [
     "derive",
     "ethereum-types",
 ] }
+reth-rlp-derive = { path = "../common/rlp-derive" }
 reth-codecs = { version = "0.1.0", path = "../storage/codecs" }
 
 # ethereum
+discv5 = { git = "https://github.com/sigp/discv5" }
 ethers-core = { git = "https://github.com/gakonst/ethers-rs", default-features = false }
 parity-scale-codec = { version = "3.2.1", features = ["derive", "bytes"] }
 tiny-keccak = { version = "2.0", features = ["keccak"] }
@@ -41,6 +43,8 @@ hex = "0.4"
 hex-literal = "0.3"
 modular-bitfield = "0.11.2"
 derive_more = "0.99"
+url = "2.3"
+generic-array = "0.14"
 
 # proof related
 triehash = "0.8"
@@ -53,6 +57,7 @@ arbitrary = { version = "1.1.7", features = ["derive"] }
 serde_json = "1.0"
 hex-literal = "0.3"
 test-fuzz = "3.0.4"
+rand = "0.8"
 
 # necessary so we don't hit a "undeclared 'std'": 
 # https://github.com/paradigmxyz/reth/pull/177#discussion_r1021172198
