+++ crates/primitives/src/lib.rs
@@ -22,6 +22,7 @@ mod hex_bytes;
 mod integer_list;
 mod jsonu256;
 mod log;
+mod net;
 mod peer;
 mod receipt;
 mod storage;
@@ -42,6 +43,7 @@ pub use hex_bytes::Bytes;
 pub use integer_list::IntegerList;
 pub use jsonu256::JsonU256;
 pub use log::Log;
+pub use net::{kad_key, NodeKey, NodeRecord, Octets};
 pub use peer::{PeerId, WithPeerId};
 pub use receipt::Receipt;
 pub use storage::StorageEntry;
